src/analytics-engine/requirements.txt
psycopg2-binary
pandas
python-dotenv
transformers
torch